+++ .eslintrc.yml
@@ -117,7 +117,7 @@ overrides:
       prettier-internal-rules/no-doc-builder-concat: error
       prettier-internal-rules/prefer-fast-path-each: error
       prettier-internal-rules/prefer-is-non-empty-array: error
-      prettier-internal-rules/no-empty-string-for-ifbreak: error
+      prettier-internal-rules/no-empty-flat-contents-for-if-break: error
   - files: src/document/*.js
     rules:
       prettier-internal-rules/no-doc-builder-concat: off
