+++ cspell.json
@@ -165,7 +165,6 @@
         "iarna",
         "ICSS",
         "idempotence",
-        "ifbreak",
         "iife",
         "IIFEs",
         "ikatyang",
